Fix Slider::test_snapMode_touch() to work for narrower sliders
When the test moved the slider to from "sliderPos" 0.0 to 0.1 (sliderPos
is in the range of [0,1]), the actual touch point movement for that
would be smaller than the drag threshold if the slider wasn't wide
enough. It turns out that for native windows style the slider width was
so narrow that this became a problem.

It did not fail for test_snapMode_mouse() because mouse doesn't have
a drag threshold for some reason.

Task-number: QTBUG-95679
Change-Id: I922d1387d3517f440a420af27fa5a8324ba88732
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit 14a61c3ea39c58da9f22a2de788ee7a34dabd240)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>