QQuickDeliveryAgentPrivate::pointerTargets: omit subscene root children
When we ask for the delivery target items in the window's contentItem,
the list should NOT include any item that is a subscene root (nor its
children), because delivery to those is done later in
QQuick3DViewport::internalPick() by calling QQuickDeliveryAgent::event()
on the subscene agent. It's premature for the main DA to deliver events
there (except in the case that it delivers an update event to a grabber,
which doesn't involve calling pointerTargets() anyway).

As for why the subscene root is a child of the window, even when it's
only being used as a texture source, that's from qtquick3d
f7f0be25cfa837758a96cdf9ef44f42401d33a9b

Fixes: QTBUG-92841
Change-Id: I6d2eb3e75b86753dfaab68afb5b99c4f1e74a473
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit a55816d076bd06320c6e372005aa1b7509a3d7eb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>