qmllint: Warn about assigning literals to read-only properties
The logic in qqmljstypepropagator.cpp only runs when there is actually
a binding on which the type propagator can run. That is not the case for
simple literal bindings, which are instead handled in
QQmlJSImportVisitor::parseLiteralBinding. So we add the check there,
too.
The formerly identical test cases for read-only properties are adjusted
so that one checks only a simple literal assignment, and the other one
only an assignment inside a script context.
Amends 0984eb6ab0db4cebdf95f17b4f81fff6a810703f.

Change-Id: Icb947c5236a30d151a0e2204de30595c2b150f8a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>