Gallery example: remove binding loop from TextArea page
The width binding for the TextArea was didn't make a lot of sense.
It used to be limited by the panes availableWidth / 3, but if the
implicitWidth was less, the text would move towards the center, since
the anchors are centering the TextArea horizontally.

Setting the width to just always be the panes availableWidth / 3,
simplifies the binding, removes the binding loop, and doesn't move the
visible text towards the center when the implicitWidth is less than 1/3
of the panes availableWidth.

Note that this doesn't directly fix QTBUG-91109, because QTBUG-91109
isn't really a bug to begin with.

Pick-to: 6.2
Fixes: QTBUG-91109
Change-Id: I5b6de8dfd11fd726e8ffdae0e46143c448985dc4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>