Ensure qmldir is written before finalizers run
Deferred CMake calls are executed in the order they are created.
When using CMake 3.19 or later, we use a deferred operation to write
the qmldir file. In static builds, qt6_import_qml_plugins() is run as
a target finalizer, but it needs the qmldir files to exist at that
point. Block finalization until after we've completed our deferred
write of the qmldir file.

Fixes: QTBUG-96290
Pick-to: 6.2
Change-Id: I6cc5cd183caa4a9c603d4549794b8f00e4358fd8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>