Fix ScrollBar::test_snapMode_touch() autotest
It assumed the travel range of the handle was the whole width of the
ScrollBar control. But this is not the case for e.g. native Windows
style, since it has decrease/increase arrow buttons on each side of the
groove.

Task-number: QTBUG-95679
Change-Id: I2f38d8186ded7710d6508ae0fe0aa2ec7ea84a13
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 44801f65fa7e1f6ecfcbde063a646df13c31c7fc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>