Reduce makeCurrents with unnecessarily different surfaces
Move the calls to explicitly make the OpenGL context current (if there
is one at all) to be within a beginFrame-endFrame block whenever
possible.

Once a frame is being recorded, makeThreadLocalNativeContextCurrent
means calling context->makeCurrent with the target window. Outside, a
QOffscreenSurface is used because no particular window is targeted at
that point from QRhi's perspective.

Now, within both render loops there are places where we do a
makeThreadLocalNativeContextCurrent unnecessarily early, before
calling beginFrame. This serves no practical purpose, but on the other
hand it gives calls like the following on the QOpenGLContext:

makeCurrent(offscreen_surface) // why is this not window already?
makeCurrent(window)
swapBuffers(window)

The purpose of this all is to ensure that applications' direct OpenGL
rendering connected to a QQuickWindow signal or issued during an
updatePaintNode have a context current on the thread. This is what Qt
5 guaranteed and Qt 6 must continue to uphold the contract, even if in
Qt 6 all the context and OpenGL stuff is hidden behind an additional
layer, and a QRhi API call gives no guarantees what it does regarding
the current context and what will be current when it returns. Hence
the need for an explicit semi-internal QRhi API call. However, this
has to follow the semantics from other aspects of the API, and so
cannot assume that it is valid to do makeCurrent with a window outside
a begin/endFrame - it does not even know the window before a
beginFrame is issued with a swapchain (which in turn targets a
window).

Now, user code with direct OpenGL usage is invoked at earliest in the
syncSceneGraph() call, and that is of course inside a
begin/endFrame. So doing a makeCurrent(offscreen_surface) early on has
no purpose at all, it is rather an artifact of the original RHI port
of Qt Quick and the Qt 5 structure of the code. So we can safely just
move the makeThreadLocalNativeContextCurrent() calls a bit closer to
the sync.

The additional advantage here is that it prevents flickering during
window resizing on macOS with OpenGL. What exactly causes empty
content to be displyed upon sequences such as this one is not quite
clear, but it could be some complex interplay between making a
QOffscreenSurface current, a QWindow current, and then sometimes
skipping the swapBuffers (which can happen during a grab or due to a
variety of other reasons)

makeCurrent(offscreen_surface)
makeCurrent(window)
swapBuffers(window)
makeCurrent(offscreen_surface)
makeCurrent(window)
// no swapBuffers here
makeCurrent(offscreen_surface)
...

Removing offscreen surface from the picture seems to solve the issues.

Change-Id: I07935e30f7f962937fd4e0973b9212eb5b043938
Fixes: QTBUG-96405
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit dfe321422b179feac4c416ed3a19aff8eb0407d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>