Reduce and clean up makeThreadLocalContext calls
Move the calls to explicitly make the OpenGL context current (if there
is one at all) to be within a beginFrame-endFrame block whenever
possible.

Once a frame is being recorded, makeThreadLocalNativeContextCurrent
means (*) calling context->makeCurrent with the target window. Outside,
a QOffscreenSurface is used because no particular window is targeted at
that point from QRhi's perspective.

(*) this will no longer be true once the OpenGL backend of QRhi in
qtbase improves its logic and avoids changing the surface when not
necessary. Regardless, the cleanups here are just as applicable.

Within both render loops there are places where we do a
makeThreadLocalNativeContextCurrent unnecessarily early, before
calling beginFrame. This serves no practical purpose, but on the other
hand it gives calls like the following on the QOpenGLContext:

makeCurrent(offscreen_surface) // why is this not window already?
makeCurrent(window)
swapBuffers(window)

The purpose of this all is to ensure that applications' direct OpenGL
rendering connected to a QQuickWindow signal or issued during an
updatePaintNode have a context current on the thread. This is what Qt
5 guaranteed and Qt 6 must continue to uphold the contract, even if in
Qt 6 all the context and OpenGL stuff is hidden behind an additional
layer, and a QRhi API call gives no guarantees what it does regarding
the current context and what will be current when it returns. Hence
the need for an explicit semi-internal QRhi API call. However, this
has to follow the semantics from other aspects of the API, and so
cannot assume that it is valid to do makeCurrent with a window outside
a begin/endFrame - it does not even know the window before a
beginFrame is issued with a swapchain (which in turn targets a
window).

User code with direct OpenGL usage is invoked at earliest in the
syncSceneGraph() call, and that is of course inside a
begin/endFrame. So doing a makeCurrent(offscreen_surface) early on has
no purpose at all, it is rather an artifact of the original RHI port
of Qt Quick and the Qt 5 structure of the code. So we can safely just
move the makeThreadLocalNativeContextCurrent() calls a bit closer to
the sync.

In addition, this allows us to clean up some leftovers from the original
6.0 rhi port. (variables like 'current' are the legacy of old OpenGL
code current = makeCurrent() and is not quite applicable the same way,
massage these as appropriate)

Change-Id: I62e0e0c211650b69ba1228058f9b3e182e968a9d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit f5dad3363067512ae798505cb7b67ae351ca3191)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>