Fix ScrollBar::test_snapMode_mouse() autotest
It assumed the travel range of the handle was the whole width of the
ScrollBar control. But this is not the case for e.g. native Windows
style, since it has decrease/increase arrow buttons on each side of the
groove.

Task-number: QTBUG-95679
Change-Id: I5edc7410d20e409f11c216515560f5edf902d901
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit b621988d76203ff42fdd70d7ab58d8c345d8a3ad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>