Ensure init of m_current_projection_matrix in single-clipped-item scene
If there is only one item in the scene, based on QSGRenderNode, with
clipping turned on, then we did not end up in render(Un)MergedBatch(),
so the usual initialization was not done, and the custom item was
effectively invisible. This is because of directly calling
renderRenderNode() from renderBatches(), bypassing the other places
where it would normally be done; so now we initialize it only in that
case in renderBatches().

Fixes: QTBUG-95881
Change-Id: Ic306ae16690103ad65a2e2ccc2f58907577e5b1b
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>