qmldom: improve standalone reparsing of bindings
Bindings can be expressions, function expressions, or blocks of
statements.
This makes using a js parser without knowing if you expect an
expression or a statement difficult.
The previous solution (wrapping the code in a function declaration)
does not work for object and (anonymous) function declarations.
We give up trying to use a plain js parser, and simply parse a binding
in a dummy object in qml mode.

Change-Id: I69e3a05568ba6cf1ec76abe0daeaf0cb1fe01ca6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>