Get rid of renderpass descriptor objects from the pipeline cache keys
Follow what we did to the srb: instead of storing an object reference,
store an opaque blob to test for compatibility. This works even if the
object is destroyed in the meantime. Thus, the annoying and error-prone
invalidatePipelineCacheDependency() goes away, finally.

All this works because creating a QRhiGraphicsPipeline with a given
QRhiRenderPassDescriptor only needs the renderpass object for building
(as in create()). Afterwards the object can be destroyed, even. The
pipeline is still usable with any render targets that have a
_compatible_ rp descriptor. Therefore, it is ideal if there is a way
to test for this compatibility without having to have the original
object around for ever.

Relies on enablers (most importantly,
QRhiRenderPassDescriptor::serializedFormat()) added to QRhi in qtbase.

While not necessarily obvious, the real benefit here is significant
performance gains in certain situations: consider rendering a scene to a
given render target, then to another one, then yet another one, for
example via QQuickRenderControl. We no longer have to do aggressive
purging of the pipelines just because a QRhiRenderPassDescriptor some of
the recently created pipelines may be associated with is going to
disappear. Rather, everything can stay as-is, there are no crashes due
to dangling pointers in the pipeline cache keys. The majority
of cases involve highly compatible render targets (e.g. rendering
into textures of the same format), so reuse is highly beneficial.

Change-Id: I8ad4bfcb3c7e66c7364c91a749e2a2e6837b4f1a
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 5e3de6aa6083f0cf312e60dc4fc9a7d999bc6c90)