Fix ScrollBar::test_touch() autotest
It assumed the travel range of the handle was the whole width of the
ScrollBar control. But this is not the case for e.g. native Windows
style, since it has decrease/increase arrow buttons on each side of the
groove.

Task-number: QTBUG-95679
Change-Id: I3cb8b8d5d91813a865eda0fc5025b8799ef32251
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 0cafcc269a0b4f6b22a75f9c946237a286899dd7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>