Turn qml, qmlscene and qmlpreview into apps
Before, they were tools, which means they are not built when
cross-compiling.  However, all these tools are likely to be desired on a
target to run QML-only projects.

The earlier way to build these apps for a cross target was to set
QT_BUILD_TOOLS_WHEN_CROSSCOMPILING to ON.  That seems like overkill.

Pick-to: 6.2
Task-number: QTBUG-96454
Change-Id: Ibbdc4e4599b0c798ddbd96804a0bbb821f546415
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>