Doc: Repair broken links to the TextMetrics Qml type
The `FontMetrics` Qml type documentation links to the tighly related
`TextMetrics` Qml type.

The links were spelled as pointing to instantiating class
`QQuickTextmetrics` instead of the Qml type `TextMetrics` in `QtQuick`,
resulting in a link to a non-existing page.

The relevant links were updated to point to `QtQuick::TextMetrics`
resolving the linking issue.

Furthermore, all the links were of the form `\l{link to some property X of
TextMetrics}{...} property of \l {link the same property
X}{TextMetrics}`.

The second link in the pattern points to the same property as the link
right before it but shows a link-text that expresses that it should
point to the page of the whole type.

Both to adjust the latter links and to simplify them, the pattern was
changed to `\l{link to some property X of TextMetrics}{...} property of
TextMetrics`, leaving the second link in the hand of QDoc's autolinker.

Task-number: QTBUG-96127
Change-Id: I0e5c325d55193baa5a6bc3c6fae021d41fb32f3f
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 5eee691d93e06563a591c63c2fae3513e2493eb7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>