TextEditor: Clean up example and use quick dialogs
The text editor example is dependent on Qt.labs for its dialogs and
native menus.

The long term goal is to remove the Qt.labs dependency, and replace it
with components from QtQuick.

Currently we have a file and a font dialog, that should be used instead
of the Qt.labs dialogs.

I'm also replacing some of the properties in the DocumentHandler, since
I don't think it makes sense for the font to be limited to only a few
font weights and styles (which can be selected through the font dialog).
Selecting the most common font weight and style is still possible
using the bold and italic shortcuts and toolbar buttons.

Some more features includes:
* An additional toolbar button, which can be used to strikeout the
  selected font. This required adding an additional glyph to
  fontello.ttf.
* When coupled with the fix for QTBUG-96934, the font for the current
  selection should automatically update the listviews in the font dialog
  to select the correct family, style, sizes and effects for what is
  selected in the document handler.

Fixes: QTBUG-95976
Change-Id: I2907a2270e2a139b1ccb3fcb3ce3a788306a42e3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit cba55a2f08b59a378e721077d03e871f291301fb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>