CMake: Fix warning about ShaderTools not being found when building examples
When building examples in-tree in a prefix build, we set QT_NO_CREATE_TARGETS
to TRUE and that causes find_package(QtFoo) calls in example projects
not to define targets.

If the targets were not found previously as part of the find_package calls
in qtdeclarative/CMakeLists.txt, this generates warnings like

 CMake Warning at lib/cmake/Qt6/Qt6Config.cmake:176 (find_package):
   Found package configuration file:
     lib/cmake/Qt6ShaderTools/Qt6ShaderToolsConfig.cmake
   but it set Qt6ShaderTools_FOUND to FALSE so package "Qt6ShaderTools" is
   considered to be NOT FOUND.  Reason given by package:
   Target "Qt6::ShaderTools" was not found.  Possibly due to
   QT_NO_CREATE_TARGETS being set to TRUE and thus
    lib/cmake/Qt6ShaderTools/Qt6ShaderToolsTargets.cmake
   was not included to define the target.

To avoid the warnings, search for the target ShaderTools package as an
optional component.

Amends 0d517b553bb8e94262d54439d3e16086896c1b7d

Pick-to: 6.2
Fixes: QTBUG-96358
Change-Id: I50c2f1516ca26c099559001e74b7e25c625be6e6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>