QuickFontDialog: make setCurrentFont() update listview indexes
For the non-native implementation of the QQuickFontDialog, when the qml
user assigns a new font to the currentFont property, the list view
indexes should be changed to point the family, style and pointSize of the
new font.

In cases where the family, style or size of the new font exists in
the model for the representing listview, the index for that listview should
be updated to the new value, or 0 if the value isn't found, or -1 if the model
is empty.

Fixes: QTBUG-96934
Change-Id: Ia7bb11d01c4a0807d0df742f34bd8aefb00104f6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3c15f90fc6b4f2b5689eaa8b70fb5ca76237540d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>