builtins.qmltypes: Fix inheritence and extension of basic types
There is no "number" type in QML. It existed in builtins.qmltypes
because we wanted to emulate the JavaScript number type. Now that we
have the real thing, we don't need to do this anymore. Furthermore, we
can declare JavaScript extensions to bool and url.

This way we get rid of the duplicate "number" type.

Change-Id: I2aea9795d0436e696483135f1b6c90f2ef997ea6
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
(cherry-picked from commit a9a17104b554044e178578e0141b6ebef6da07cc)
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>