QDeferredSharedPointer: fix the build with GCC 12
Comparing via operator== a QDeferredSharedPointer with a QSharedPointer
is ambiguous, as both convert to each other implictly, and there's only
operator==(QDSP, QDSP) and operator==(QSP, QSP) defined. Therefore a
"mixed" comparison is ambiguous: there's no dedicated operator==,
and one can convert either parameter to the "other type" to get a
viable candidate, and no conversion is better than the other.
The result is:

../src/qmlcompiler/qqmljstyperesolver.cpp: In function ‘bool isAssignedToDefaultProperty(const ConstPtr&, const ConstPtr&)’:
../src/qmlcompiler/qqmljstyperesolver.cpp:565:29: error: ambiguous overload for ‘operator==’ (operand types are ‘QSharedPointer<const QQmlJSScope>’ and ‘const ConstPtr’ {aka ‘const QDeferredSharedPointer<const QQmlJSScope>’})
  565 |         if (binding.value() == child)
      |             ~~~~~~~~~~~~~~~ ^~ ~~~~~
      |                          |     |
      |                          |     const ConstPtr {aka const QDeferredSharedPointer<const QQmlJSScope>}
      |                          QSharedPointer<const QQmlJSScope>

Given that apparently implicit convertibility between QDSP and QSP
is wanted, add mixed comparison operators to remove the ambiguity.

Pick-to: 6.2
Change-Id: Ifcb1a2d86986d83c6f488097c9b4c8cb35db31d9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>