QQuickDragEvent: Use proper types instead of QQmlV4Function
With QQmlV4Function we have no information at compile time about what
a arguments a function expects, and what it returns. Considering that
getDataAsString does not actually have an overload set, but simply
expects and returns a string, we can change to QString types. For
non-QString parameters, the engine will then transparently take care of
converting them.
The same applies to getDataAsArrayBuffer and QByteArray/ArrayBuffer.

Change-Id: I007796c3c01a6f021babd0dff2832506c7c63c30
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>