Doc: Repair broken link to the FlowerPower demo
The `Context2d QML Type` documentation links to the FlowerPower demo as
an example of the use of `bezierCurveTo`.

The site hosting the demo seems to have been down for a long time.
To avoid losing the chance to show such an interesting demo, the link
was replaced with an equivalent one from the Internet Archive's Wayback
Machine.

Task-number: QTBUG-96127
Pick-to: 6.2
Change-Id: I1c88776b7d689d4abd6602dd2f928f8bca602689
Reviewed-by: Paul Wicking <paul.wicking@qt.io>