CMake: Walk plugin private deps for computed static qml plugin deps
When setting up extra dependencies for a qml static plugin we only
walked private dependencies of the backing library, but the plugin
itself might have extra dependencies on it which are not on the
backing library.

An example of this is the qtquickcontrols2plugin which links
against QuickControls2ImplPrivate.

This means the qtquickcontrols2impl plugin should be a dependency of
the qtquickcontrols2plugin.

Make sure to walk the plugin's private dependencies in addition to the
backing lib private dependencies. Also make sure not to accidentally
create a linking cycle by filtering out the current plugin from being
added to itself as a result of the walk.

Amends 6fd1216801f078f65be2cbc748cc459cb6912a4f

Pick-to: 6.2
Fixes: QTBUG-97099
Change-Id: I61981826283944b160888dd64f3a6afd23495d79
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>