QQmlEngine: Fine grained translation binding tracking
This commit changes translation bindings in such a way that they always
depend on the a newly introduced translationLanguage property of the QML
engine.
Moreover, captureTranslationBinding is changed in the same way, hence
any binding containing a call to a translation function like qsTr will
also be refreshed. After those changes, QQmlEngine::retranslate can be
implemented as a notify on translationLanguage. Finally,
QQmlTranslationPropertyBinding is also changed to depend on
translationLanguage instead of uiLanguage to ensure consistency between
the various binding types.

Note that we do not use the existing uiLanguage property of QJSEngine.
If we were to do this, changing uiLanguage would already trigger the
binding reevaluation, which is a behavior change. In case someone is
manually connecting uiLanguageChanged to QQmlEngine::retranslate, this
would cause a double evaluation of translation bindings. In a future
change, we can however introduce a new enableAutoTranslationUpdate
method, which would cause binding evaluation as soon as uiLanguage is
changed.

Special care is taken for ListModel/ListElement, which does not store a
proper translation binding, but instead operates on
QV4::CompiledData::Binding. For now, we simply refresh the complete
ListModel in case it contains a translation binding. It should be
possible to optimize this further in a future change.
Note that only "static" ListElement properties can contain translations:
The dynamic insertion API makes it impossible to insert a translation,
as it would already be resolved to a string when insert is called.

[ChangeLog][QQmlEngine] QQmlEngine::retranslate no longer refreshes all
bindings, but only translation bindings.

Pick-to: 6.2
Fixes: QTBUG-96192
Change-Id: I54d213fd46b6914e8f686843b49e155909117218
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>