Material: fix PageIndicator delegate creation error
We can't make "pressed" a required property, as QQuickPageIndicator
doesn't create the delegates (the repeater does), and so it can't set
it as an initial property.

This is already tested, but as mentioned in the comments of
QTBUG-96876, the error is printed in the output but doesn't affect the
test.

Fixes: QTBUG-96876
Pick-to: 6.2
Change-Id: I626148aabdc908ac6e171b9a05e65741e6dc871d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>