Material: fix PageIndicator delegate creation error
We can't make "pressed" a required property, as QQuickPageIndicator
doesn't create the delegates (the repeater does), and so it can't set
it as an initial property.

This is already tested, but as mentioned in the comments of
QTBUG-96876, the error is printed in the output but doesn't affect the
test.

Fixes: QTBUG-96876
Change-Id: I626148aabdc908ac6e171b9a05e65741e6dc871d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 852c8d1d8a953c626735f044838454a4d249c225)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>