QQmlDelegateModelAttached: Make static properties statically known
We know that the attached properties for persistedItems and items always
exist, so we can add them statically to the attached property. That way,
it is possible to define them in a declarative way/assign bindings to
them.
We still need to insert the static properties in the
QQmlDelegateModelItemMetaType, as not every code path accessing them
creates the attached object; see tst_qquickvisualdatamodel::insert,
especially the tests involving Package.

As a byproduct of this change, we now create a valid attached property
in more cases. However, the fact that the properties are only accessible
from the toplevel object of the delegate was never documented anywhere,
and we still should not create attached properties unless they are
explicitly requested.

[ChangeLog][QtQml.Models][DelegateModel] DelegateModel's attached
properties are now valid everywhere in its delegate, not only in the
toplevel object.

Note that this is impossible to achieve with the dynamically created
properties that get defined according to the groups values.

Task-number: QTBUG-86045
Change-Id: I55519b44973f4b50aa57fad418ec606dc4fe12be
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>