Revert "Fix for possible crash in QSGDefaultLayer::grab"
This reverts commit 1c5de027d0c31d1d6697bd0557128d92207763d8.

The fix here is not correct. Calling a QSGRhiLayer function from the gui
thread is very wrong and can cause a set of unexpected issues. The
Address Sanitizer catches this by recognizing that the render thread is
trying to do something with an object destroyed in the meantime on the
main thread in the layer->setItem(null) call.

The issue the original fix is trying to address needs to be addressed in
some different form.

Fixes: QTBUG-94975
Pick-to: 6.2 6.1 5.15
Change-Id: I46f904026281201fc6d233ed7d3bdc7080934afe
Reviewed-by: Christian Strømme <christian.stromme@qt.io>