StackView: complete animations when using StackView.Immediate transition
The idea behind using StackView.Immediate is to immediately move the
stack view item into the new state without running any animations.
However the preexisting code in QQuickStackViewPrivate::completeTransition
was doing nothing, because the animations didn't have the proper target set.
So they could never be started and/or completed.
As a result, the pushed stack view item could appear at the wrong position
and with the other properties set incorrectly (for example, opacity).

The new code uses the same approach, as in startTransition(), but simplifies
it to execute only the animations.
Instead of actually running the animations, a new
QAbstractAnimationJob::complete() method is added, which simulates the
animation progress by starting the animation, forwarding it to the end, and
completing it. This results in all the properties receiving the correct
values, and the stack view items being shown correctly.

Fixes: QTBUG-96966
Fixes: QTBUG-61496
Change-Id: I990a133881c66e3ecb83887c60596a5d45e570d9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 5339b3eb13f4a175372d46d91e2fc0be3c145344)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>