Fix distorted subpixel text for non-RHI code path
This amends 4757cac470edbeaeaceca4e63075d9f1139f546b which is a
cherry-pick of the fix of subpixel positioning of text for the
RHI code path. But this code path is not enabled by default in Qt 5,
so in order to fix this properly, we need to make the equivalent
changes to the runtime-compiled shaders.

Task-number: QTBUG-96112
Task-number: QTBUG-83626
Change-Id: I6c104fd74298e1ec503b9e84b8cc884fb78ddb7b
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>