Fix tst_qquickwindow::grab with non-integer dpr
Do not introduce yet another way of calculating the scaled size.
Truncating the width * dpr and height * dpr is separately not the
same as doing size * dpr (because that rounds). And it is the
latter that is used everywhere.

This makes the case not fail anymore with device pixel ratios such
as 1.25.

Pick-to: 6.2
Change-Id: Id5d7b1799977967b3110925bf7ccd0417c6bdaa9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>