Protect against cycles when iterating types and extensions
We do check for cycles in endVisit(UiProgram*), but we may try to
iterate the base types already before. In that case, a cycle sends us
into an infinite loop.

The extra duplicate trackers are necessary because 6.2 does not have the
generic type inference, yet, and it still does the inheritance check
after a number of other checks.

Change-Id: I629beb873a6ac8b15570eadce0bb1b8ce70ae266
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 26da93b062d6162a8275fb6535b422db9716730b)